requirements.txt
Flask
requests
python-dotenv
bakong-khqr
qrcode[pil]
gunicorn
